AIVA Full Stack Developer AI Applications (Python + React)/Assessment 1/crm-hcp-module/backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
python-dotenv>=1.0.0
pydantic>=2.5.0
langchain>=0.1.0
langgraph>=0.0.9
langchain-groq>=0.1.0
aiohttp>=3.9.0
requests>=2.31.0
python-multipart>=0.0.6
numpy>=2.0.0
sqlalchemy>=2.0.0
pymysql>=1.1.0
mysql-connector-python>=9.0.0

fastapi
uvicorn
pydantic
langgraph
langchain-groq
langchain-core
python-dotenv
sqlalchemy
pymysql
mysql-connector-python